Removal of inner classes on mixins, factor them out
